Adds a vector tile layer for the map

Adds a vector tile layer for the map

geom_vector_tiles(
  ...,
  type = c("water", "buildings", "places", "transit", "pois", "boundaries", "roads",
    "earth", "landuse"),
  nextzen_api_key = getOption("nextzen_API_key"),
  tile_size_px = NULL,
  transform = function(d) d
)

Arguments

...
extra arguments

type
type of vector layer

nextzen_api_key
nextzen API key for vector tile data

tile_size_px
tile size in pixels, may increase resolution of tile data

transform
transform function to apply to the vector tile data

Value

a geom_sf object