Cut values with pretty labels

Cut values with pretty labels

pretty_cut(
  values,
  breaks,
  format = function(d) d,
  binding = "to",
  spacing = " ",
  under_text = "<",
  over_text = ">",
  ...
)

Arguments

values
a numeric vector which is to be converted to a factor by cutting.

breaks
a numeric vector of two or more unique cut points

format
a function specifying how to format values in the labels

binding
a string binding two values in the label

spacing
a spacing string between elements in the label

under_text
a string taking the space of the lower value and binding if the lower value is `-Inf`

over_text
a string taking the space of the upper value and binding if the upper value is `-Inf`

...
additional arguments passed to `cut`

Value

object, (potentially cached version)